InputProcessor semantic changes.
InputProcessor now uses first match instead of finding the longest match.
Fixed possible bug where keycombo did not match because there were extra
bytes on input.

+++ cppconsui/InputProcessor.h
@@ -177,7 +177,7 @@ class InputProcessor
 		 *		- -needed where needed is how many chars are needed for a match
 		 *		- 0 if they not match
 		 */
-		int Match(const std::string &skey, const char *input, const size_t bytes);
+		int Match(const std::string &skey, const char *input, const int bytes);
 
 		/** adds the bindable to the keymap if it has a nonempty keyvalue defined */
 		void MapBindable(const Bindable& bindable);
